(MAINT) Set GHA to use hugo-module-update action
Prior to this commit the workflow for keeping hugo module dependencies
up to date included custom scripting in this repository.

This commit updates the workflow defined in #16 to use the newly
published platenio/action-hugo-module-update, reducing the code needed
to keep the dependencies up to date and improving reusability for this
project and projects which are built from this template.