(MAINT) Convert to using new Platen
This change updates the template to point to the new, canonical project
for Platen where all future development is being done: platenio/platen.

This represents a major break from the prior version, which consumed the
hugo-book theme as a basis. Platen has been extensively rewritten based
on the theme but is no longer intended to keep in parity with it.

This change focuses on the user experience and especially the
flexibility and extensibility of owning the module top to bottom.

In the future, further changes and documentation will be needed to
make using this project as delightful as possible.